+++ config/thumbnailable.php
@@ -2,91 +2,59 @@
 
 return array(
 
-	// where to store the image files
-	// (you'll need to create this directory)
-	'storage_dir'   => path('storage'). 'uploads' . DS . 'thumbnails',
+	/**
+	 * Where do we store the image files?
+	 * (we'll try and create this directory if it doesn't exist)
+	 */
+	'storage_dir' => path('storage') . 'uploads' . DS . 'thumbnails',
 
 
-	// Should we keep the original uploaded file?
+	/**
+	 * Should we keep the original uploaded file?
+	 * This needs to be true if you set `strict_sizes` to be false.
+	 */
 	'keep_original' => true,
 
 
-	// If this is true, you can only request thumbnails of the
-	// sizes defined below.
-	// If this is false, then you can request a thumbnail of any
-	// size (this requires that "keep_original" is true.)
-	'strict_sizes'  => true,
+	/**
+	 * If this is true, you can only request thumbnails of the
+	 * sizes defined below.
+	 * If this is false, then you can request a thumbnail of any
+	 * size (this requires that `keep_original` above is true.)
+	 */
+	'strict_sizes' => true,
 
 
-	// If this is true, the resized images are generated when
-	// the model is saved.  If this is false, they are generated
-	// only when requested (and then cached).
-	'on_save'       => true,
+	/**
+	 * If this is true, the resized images are generated when
+	 * the model is saved.  If this is false, they are generated
+	 * on-demand (and then cached).
+	 */
+	'on_save' => true,
 
 
-	// How to resize the images:
-	// crop - fit the shortest side into the formatted dimension, keep aspect ratio
-	// fit - fit the longest side into the formatted dimensions, keep aspect ratio
-	// resize - resize image to new dimensions, aspect ratio be damned!
+	/**
+	 * How should we resize the images (value passed to the Resizer bundle).
+	 * Set to one of the following strings:
+	 * "crop"  - resizes the image so that it completely fills the thumbnail,
+	 *           but maintains the original image's aspect ratio.
+	 * "fit"   - resizes the image so that it fits enitrely inside the thumbnail
+	 *           (padding it, if necessary) and maintains the aspect ratio.
+	 * "exact" - simply resizes the original image to the dimensions of the
+	 *           thumbnail with no regard for aspect ratio.
+ 	 */
 	'resize_method' => 'crop',
 
 
-	// Image format for thumbnailed images.
-	// Can be 'jpg', 'png' or 'gif'
+	/**
+	 *  Image format for thumbnailed images: one of "jpg", "png" or "gif".
+	 */
 	'thumbnail_format' => 'png',
 
 
-	// Image quality for thumbnails.
-	// 0-100.  Irrelevant if format is "gif"
+	/**
+	 *  Image quality for thumbnails: 0-100.  Irrelevant if format is "gif"
+	 */
 	'thumbnail_quality' => 80,
-
-
-	// For thumbnails that including padding, what color should the padded
-	// area in the resulting thumbnail be.  If it's not set, or set to null,
-	// and the format is "png", then it will be transparent.
-	// Should be an array of (r,g,b) values
-	'thumbnail_background' => array( 255, 255, 255),
-
-
-	// The array of models that are thumbnailable and their configurations.
-	// These can be defined in the main application/config/thumbnailable.php file,
-	// or defined within the models as:
-	//
-	//   public static $thumbnailable = array(...)
-	//
-
-	/* EXAMPLE
-
-	'user' => array(
-
-		// what image to return if none is given
-		'default_field' => 'image',
-
-		// list of fields in the model that are thumbnailable
-		'fields'        => array(
-
-			'image' => array(
-
-				// sizes
-				// - An array where the key is the "nickname" for the format and the value is
-				//   a "(height)x(width)" dimension for the thumbnail.
-				'sizes' => array(
-					'small'   => '50x50',
-					'medium'  => '100x100',
-					'large'   => '300x300',
-				),
-
-				// what size to return if none is given
-				'default_size' => 'small',
-
-				// other possible keys can override the base settings for this model, i.e.:
-				// - storage_dir
-				// - keep_original
-
-			)
-		)
-
-	)
-	*/
 
 );